+++ src/main/java/com/cricketcraft/cavegame/client/rendering/SetBlocks.java
@@ -1,0 +1,9 @@
+package com.cricketcraft.cavegame.client.rendering;
+
+public class SetBlocks
+{
+    public static void SetBlocks()
+    {
+        //BlockInformation.AllBlocks;
+    }
+}
