+++ src/main/java/com/cricketcraft/cavegame/core/item/WeaponMelee.java
@@ -1,0 +1,6 @@
+package com.cricketcraft.cavegame.core.item;
+
+public class WeaponMelee
+{
+
+}
