+++ src/main/java/com/cricketcraft/cavegame/core/util/constants/BlockInformation.java
@@ -1,0 +1,9 @@
+package com.cricketcraft.cavegame.core.util.constants;
+
+import com.cricketcraft.cavegame.core.world.Chunk;
+
+public class BlockInformation
+{
+    public static final int TOTAL_CHUNKS = 5000;
+    public static Chunk[][] ChunkInfo = new Chunk[TOTAL_CHUNKS][TOTAL_CHUNKS];
+}
