+++ src/main/java/com/cricketcraft/cavegame/core/util/constants/TexturePrefix.java
@@ -1,0 +1,6 @@
+package com.cricketcraft.cavegame.core.util.constants;
+
+public class TexturePrefix {
+    public static final String block = "assets/textures/blocks/";
+    public static final String item = "assets/textures/items";
+}
