--- src/main/java/com/cricketcraft/client/rendering/SetBlocks.java
@@ -1,9 +1,0 @@
-package com.cricketcraft.client.rendering;
-
-public class SetBlocks
-{
-    public static void SetBlocks()
-    {
-        //BlockInformation.AllBlocks;
-    }
-}
