--- src/main/java/com/cricketcraft/core/items/WeaponMelee.java
@@ -1,6 +1,0 @@
-package com.cricketcraft.core.items;
-
-public class WeaponMelee
-{
-
-}
