--- src/main/java/com/cricketcraft/core/utils/constants/BlockInformation.java
@@ -1,11 +1,0 @@
-package com.cricketcraft.core.utils.constants;
-
-import com.cricketcraft.core.core.Chunk;
-import java.util.ArrayList;
-import java.util.List;
-
-public class BlockInformation
-{
-    public static final int TOTAL_CHUNKS = 5000;
-    public static Chunk[][] ChunkInfo = new Chunk[TOTAL_CHUNKS][TOTAL_CHUNKS];
-}
