--- src/main/java/com/cricketcraft/core/utils/constants/TexturePrefix.java
@@ -1,6 +1,0 @@
-package com.cricketcraft.core.utils.constants;
-
-public class TexturePrefix {
-    public static final String block = "assets/textures/blocks/";
-    public static final String item = "assets/textures/items";
-}
