+++ src/main/resources/com/cricketcraft/cavegame/assets/textures/blocks/stone.png
